Sort out package imports and structure
