+++ agent/deployment_stages/common.py
@@ -1,7 +1,7 @@
 # Copyright (c) Trainline Limited, 2016-2017. All rights reserved. See LICENSE.txt in the project root for license information.
 
 import os, sys, yaml
-from deployment_scripts import PowershellScript, ShellScript
+from .deployment_scripts import PowershellScript, ShellScript
 
 class DeploymentError(RuntimeError):
     pass
