+++ agent/deployment_stages/register_with_consul.py
@@ -1,6 +1,6 @@
 # Copyright (c) Trainline Limited, 2016-2017. All rights reserved. See LICENSE.txt in the project root for license information.
 
-from common import *
+from .common import DeploymentStage
 
 class RegisterWithConsul(DeploymentStage):
     def __init__(self):
