+++ agent/deployment_stages/stop_application.py
@@ -1,6 +1,7 @@
 # Copyright (c) Trainline Limited, 2016-2017. All rights reserved. See LICENSE.txt in the project root for license information.
 
-from common import *
+import os
+from .common import LifecycleHookExecutionStage, get_previous_deployment_appspec
 
 class StopApplication(LifecycleHookExecutionStage):
     def __init__(self):
